broken component into two
